% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{BFI2}
\alias{BFI2}
\title{BFI2 data}
\format{
An object of class \code{data.frame} with 838 rows and 60 columns.
}
\source{
\url{https://osf.io/awvnd}
}
\usage{
data(BFI2)
}
\description{
Data obtained from the Big Five Inventory-2 (BFI-2; Soto & John, 2017), collected from adolescents ages 14 to 17 years old enrolled in a high school AP Statistics course (Ober et al., 2021). 
The inventory contains 60 items on a 5-point Likert scale, where exactly 12 items load onto each of the five personality factors in the BFI-2 (Extraversion, Agreeableness, Conscientiousness, Negative Emotionality, and Open-Mindedness).  
Data for 838 subjects is contained, after removing four subjects with missing data. No missing data remains. 
Variables are labeled by the first letter of the personality factor (e.g., E = Extraversion, A = Agreeableness, etc.) and item number 1 through 12.
Items marked with an 'R' are negatively worded and were properly reverse coded for this dataset.

\itemize{
  \item{\code{E1}}{ Is outgoing, sociable.}.
  \item{\code{A1}}{ Is compassionate.}.
  \item{\code{C1R}}{ Tends to be disorganized. [R]}.
  \item{\code{N1R}}{ Is relaxed, handles stress well. [R]}.
  \item{\code{O1R}}{ Has few artistic interests. [R]}.
  \item{\code{E2}}{ Has an assertive personality.}.
  \item{\code{A2}}{ Treats others with respect.}.
  \item{\code{C2R}}{ Tends to be lazy. [R]}.
  \item{\code{N2R}}{ Stays optimistic after experiencing a setback. [R]}.
  \item{\code{O2}}{ Is curious about many different things.}.
  \item{\code{E3R}}{ Rarely feels excited or eager. [R]}.
  \item{\code{A3R}}{ Tends to find fault with others. [R]}.
  \item{\code{C3}}{ Is dependable, steady.}.
  \item{\code{N3}}{ Is moody, has up and down mood swings.}.
  \item{\code{O3}}{ Is inventive, finds clever ways to do things.}.
  \item{\code{E4R}}{ Tends to be quiet. [R]}.
  \item{\code{A4R}}{ Feels little sympathy for others. [R]}.
  \item{\code{C4}}{ Is systematic, likes to keep things in order.}.
  \item{\code{N4}}{ Can be tense.}.
  \item{\code{O4}}{ Is fascinated by art, music, or literature.}.
  \item{\code{E5}}{ Is dominant, acts as a leader.}.
  \item{\code{A5R}}{ Starts arguments with others. [R]}.
  \item{\code{C5R}}{ Has difficulty getting started on tasks. [R]}.
  \item{\code{N5R}}{ Feels secure, comfortable with self. [R]}.
  \item{\code{O5R}}{ Avoids intellectual, philosophical discussions. [R]}.
  \item{\code{E6R}}{ Is less active than other people. [R]}.
  \item{\code{A6}}{ Has a forgiving nature.}.
  \item{\code{C6R}}{ Can be somewhat careless. [R]}.
  \item{\code{N6R}}{ Is emotionally stable, not easily upset. [R]}.
  \item{\code{O6R}}{ Has little creativity. [R]}.
  \item{\code{E7R}}{ Is sometimes shy, introverted. [R]}.
  \item{\code{A7}}{ Is helpful and unselfish with others.}.
  \item{\code{C7}}{ Keeps things neat and tidy.}.
  \item{\code{N7}}{ Worries a lot.}.
  \item{\code{O7}}{ Values art and beauty.}.
  \item{\code{E8R}}{ Finds it hard to influence people. [R]}.
  \item{\code{A8R}}{ Is sometimes rude to others. [R]}.
  \item{\code{C8}}{ Is efficient, gets things done.}.
  \item{\code{N8}}{ Often feels sad.}.
  \item{\code{O8}}{ Is complex, a deep thinker.}.
  \item{\code{E9}}{ Is full of energy.}.
  \item{\code{A9R}}{ Is suspicious of others' intentions. [R]}.
  \item{\code{C9}}{ Is reliable, can always be counted on.}.
  \item{\code{N9R}}{ Keeps their emotions under control. [R]}.
  \item{\code{O9R}}{ Has difficulty imagining things. [R]}.
  \item{\code{E10}}{ Is talkative.}.
  \item{\code{A10R}}{ Can be cold and uncaring. [R]}.
  \item{\code{C10R}}{ Leaves a mess, doesn't clean up. [R]}.
  \item{\code{N10R}}{ Rarely feels anxious or afraid. [R]}.
  \item{\code{O10R}}{ Thinks poetry and plays are boring. [R]}.
  \item{\code{E11R}}{ Prefers to have others take charge. [R]}.
  \item{\code{A11}}{ Is polite, courteous to others.}.
  \item{\code{C11}}{ Is persistent, works until the task is finished.}.
  \item{\code{N11}}{ Tends to feel depressed, blue.}.
  \item{\code{O11R}}{ Has little interest in abstract ideas. [R]}.
  \item{\code{E12}}{ Shows a lot of enthusiasm.}.
  \item{\code{A12}}{ Assumes the best about people.}.
  \item{\code{C12R}}{ Sometimes behaves irresponsibly. [R]}.
  \item{\code{N12}}{ Is temperamental, gets emotional easily.}.
  \item{\code{O12}}{ Is original, comes up with new ideas.}.
}
}
\details{
Polytomous data from the Big Five Inventory-2

The BFI2 is a 60‑item personality assessment measuring the Big Five domains: Agreeableness, Conscientiousness, Extraversion, Negative Emotionality, and Open‑Mindedness. 
Each domain contains 12 items, organized into 15 facets (e.g., Compassion, Organization, Anxiety). Items are scored on a 1–5 Likert scale and include both positively and negatively keyed statements.
}
\references{
Ober, T. M., Cheng, Y., Jacobucci, R., & Whitney, B. M. (2021, January). Examining the factor structure of the Big Five Inventory-2 personality domains with an adolescent sample. \emph{Psychological Assessment, 33}(1), 14–28. \url{doi:10.1037/pas0000962}

Soto, C. J., & John, O. P. (2017). The next Big Five Inventory (BFI-2): Developing and assessing a hierarchical model with 15 facets to enhance bandwidth, fidelity, and predictive power. \emph{Journal of Personality and Social Psychology, 113}, 117–143. \url{http://dx.doi.org/10.1037/pspp0000096}
}
\keyword{datasets}
