% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robust.R
\name{choose.tuco}
\alias{choose.tuco}
\title{Plot histogram of residuals along plot of the weight function}
\usage{
choose.tuco(r, H = NULL, B = NULL, x.axis = NULL)
}
\arguments{
\item{r}{A vector of residuals}

\item{H}{Huber tuning parameter}

\item{B}{Bisquare tuning parameter}

\item{x.axis}{A vector containing two values: the lower and upper limits for the x-axis. Default is the minimum and maximum values of the residuals supplied in \code{r}, excluding infinite values.}
}
\value{
Histogram plot of residuals beneath a graph of the weight functions vs. the residuals.
}
\description{
Plot a histogram of residuals along the graph of the weighting function (dependent on the tuning parameter) as a function of the residual
}
\details{
This function allows the user to visualize the amount of downweighting applied to their data in robust latent trait estimation based on the tuning parameter and item-level residuals, in order to choose a tuning parameter that suits their data well.
              The values \eqn{H=1} and \eqn{B=4} are recommended for Huber (Huber, 1981) and bisquare (Mosteller & Tukey, 1977) weighting, respectively, but the amount of downweighting can be made more conservative or more liberal depending on the research goals.
              To increase the downweighting applied in estimation, use a smaller tuning parameter. To decrease the amount of downweighting, apply a greater tuning parameter.
              The histogram of residuals is plotted below a plot of the weight curve(s) at the tuning parameter. If \emph{H} is supplied to the function, the Huber curve is plotted, and, if \emph{B} is supplied, Tukey's bisquare weight is plotted. If both \emph{H} and \emph{B} are supplied, the weights are plotted simultaneously.
              If \emph{H} is supplied, dotted vertical lines at \emph{H} and \emph{-H} reveal the cut between data that is downweighted (\eqn{|r_ij|>H}) and data that is not downweighted (\eqn{|r_ij|<H}), e.g., receives full weighting (\eqn{w_ij=1.0}).
              If \emph{B} is supplied, dotted vertical lines at \emph{B} and \emph{-B} reveal the cut between data that receives some weight (\eqn{|r_ij|<B}) and data that is removed (\eqn{|r_ij|>B}), e.g., receives a weight of 0.
              If no tuning parameter is supplied, the histogram of residuals is provided.
}
\examples{

### Visualize Huber weight function
##  Use residual() function
thetas<-c(-2, -1, 0, 1, 2)
ipars<-data.frame(a = c(1, 1.2, 1.5, 1, 1.2, 1.5, 1, 1.2, 1.5), b = c(-2, -1.5, -1, -.5, 0, .5, 1, 1.5, 2))
info_r<-residual(thetas, "2PL", ipars, resid = "information")
choose.tuco(info_r, H=1)


### Compare Huber and bisquare weights
##  Use robust.theta() output
data(BFI2)
dat<-BFI2[,20:79]
specific<- rep(1:5, 12) #which factor each item loads on
mod<-mirt(dat, specific, TOL=0.001, method="QMCEM")
ipars<-coef(mod, simplify=T)$items
ipars<-cbind(ipars[,1:5], -ipars[,6:9]/rowSums(ipars[,1:5])) # reparameterize
out <- robust.theta(dat, ipars, model= "MGRM", resid = "standardized", dimen=5, D=1)

choose.tuco(r=out$residuals_MLE, H=1, B=4, c(-6, 6))
}
\references{
Huber, P. (1981) \emph{Robust Statistics}. Wiley, New York. https://doi.org/10.1002/0471725250

Mosteller, F., & Tukey, J. W. (1977). \emph{Data Analysis and Regression: A Second Course in Statistics}. Reading, MA: Addison-Wesley Pub Co.
}
