% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robust.R
\name{hist.weights}
\alias{hist.weights}
\title{Histogram of RMMLE Weights}
\usage{
\method{hist}{weights}(dat, freq = TRUE, ...)
}
\arguments{
\item{dat}{A \eqn{N\times J} matrix of response data}

\item{freq}{Logical; if \code{TRUE}, the histogram graphic is a representation of frequencies; if \code{FALSE}, probability densities are plotted (so that the histogram has a total area of one). Defaults to \code{TRUE}.}

\item{...}{Additional arguments to be passed to \code{mirt()}}
}
\value{
List with four histograms
 \describe{
   \item{\code{Person-Fit Indices (MMLE)}}{Histogram of person-fit residuals \eqn{l_z^{`*`}} across individuals based on the IRT model estimated with MML.}
   \item{\code{Person-Fit Indices (RMMLE)}}{Histogram of person-fit residuals \eqn{l_z^{`*`}} across individuals based on the IRT model estimated with robust MML.}
   \item{\code{Subject-Level Weights (MMLE)}}{(\eqn{N \times L}) Histogram of person-level weights based on the IRT model estimated with MML.}
   \item{\code{Subject-Level Weights (RMMLE)}}{(\eqn{N \times L}) Histogram of person-level weights based on the IRT model estimated with robust MML.}
}
}
\description{
The following function returns a histogram of weights computed for robust MML estimates of item parameters (Hong & Cheng, 2019). 
Weights are derived based on the person-fit statistic \eqn{l_z^{`*`}}, where lower values indicate greater misfit between the response pattern and the model (Snijders, 2001). 
Since \eqn{l_z^{`*`}} follows the asymptotic standard normal distribution, weights are the normalized \emph{p}-value of the residual under a one-sided alternative hypothesis, e.g.,
\eqn{w_i = \frac{\Phi(l^{`*`}_{zi})}{\sum_{i=1}^N \Phi(l^{`*`}_{zi})},}
where \eqn{\Phi(\cdot)} is the CDF of the standard normal distribution.
}
\details{
The IRT model is estimated with, first, MML and, second, RMML. Four histograms are returned. Two histograms display the distribution of person-fit residuals, \eqn{l^{`*`}_{zi}}, 
         across the sample. In the first of these histograms, \eqn{l^{`*`}_{zi}} is calculated from item parameters estimated with MML, while the second histogram displays the residuals calculated using robust MML item parameter estimates.
         As RMML reduces the bias in item parameter estimates due to aberrant responding, it is expected that person-fit residuals consequently are more accurate (e.g., see ).  
         Likewise, the latter two histograms display the distribution of person-specific weights, with the first graph utilizing MML and the second graph utilizing robust MML item parameter estimates.
         Under the null hypothesis, the p-values for a continuous test statistic are said to follow a standard normal distribution. 
         Given that the weights recommended for robust MML estimation are proportional to the p-values of \eqn{l^{`*`}_{zi}} according to the standard normal distribution, it is expected that, under model fit, the weights produce an approximately uniform distribution.
         Deviations from the uniform distribution, under appropriately large sample sizes, may provide insight into the behavior of the examinees, e.g., a large uptick near 0 may indicate prevalent aberrant responding.
         As with the residuals, these weights are expected to become more accurate after robust weighting and may reveal anomalous patterns more closely.
         Note that the distribution of weights may differ from \eqn{U(0,1)} due to the inverse weighting of the p-value by the sum of weights across the sample.
}
\examples{

# SAT12 Data: 3-Dimensional MIRT Model
SAT12[SAT12 == 8] <- NA #set 8 as a missing value
# Correct answer key
library(mirt)
key <- c(1,4,5,2,3,1,2,1,3,1,2,4,2,1,5,3,4,4,1,4,3,3,4,1,3,5,1,3,1,5,4,5) 
scoredSAT12 <- key2binary(SAT12, key)
specific <- c(2, 3, 2, 3, 3, 2, 1, 2, 1, 1, 1, 3, 1, 3, 1, 2, 1, 1, 3, 3, 1, 1, 3, 1, 3, 3, 1, 3, 2, 3, 1,2) #which factor each item loads on

hist_out <- hist.weights(scoredSAT12, model=specific)

# Compare weights before and after robust estimation
hist_out$`Subject-Level Weights (MMLE)` 
hist_out$`Subject-Level Weights (RMMLE)`

}
\references{
Hong, M., & Cheng, Y. (2019). Robust maximum marginal likelihood (RMML) estimation for item response theory models. Behavior Research Methods, 51(2), 573–588. https://doi.org/10.3758/s13428-018-1150-4

Snijders, T. A. (2001). Asymptotic null distribution of person fit statistics with estimated person parameter. Psychometrika, 66(3), 331-342.

Sinharay, S. (2016). The choice of the ability estimate with asymptotically correct standardized person-fit statistics. British Journal of Mathematical and Statistical Psychology, 69(2), 175–193. https://doi.org/10.1111/bmsp.12067.
}
