% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robust.R
\name{huber}
\alias{huber}
\title{Huber Weighting Function}
\usage{
huber(r, H)
}
\arguments{
\item{r}{A residual that measures the inconsistency of a response from the subject's assumed response model, on one item. Residuals of value NA are given a weight of 0.}

\item{H}{Huber tuning parameter. Larger values lead to less downweighting. For robust estimation, \code{H} is often set to 1.0 (Filonczuk & Cheng, 2025; Schuster & Yuan, 2011).}
}
\value{
Huber weight value.
}
\description{
Calculate the Huber weight (Huber, 1981) given a residual and Huber tuning parameter.
}
\examples{

# 1-person case
r <- c(-3, -1, 0, 1.5, 4) 
H <- 1.5 
huber(r, H)

# multi-person, multi-item case
r <- matrix(c( -2, -0.5, 0, 2, 
                3, 1.2, -1, 0.3 ), 
                nrow = 2, byrow = TRUE) 
H <- 1 
huber(r, H)
}
\references{
Filonczuk, A., & Cheng, Y. (2025). Robust estimation of the latent trait in graded response models. Behavior Research Methods, 57(1), 55. https://doi.org/10.3758/s13428-024-02574-2

Huber, P. (1981) \emph{Robust Statistics}. Wiley, New York. https://doi.org/10.1002/0471725250.

Schuster, C., & Yuan, K.-H. (2011). Robust Estimation of Latent Ability in Item Response Models. \emph{Journal of Educational and Behavioral Statistics}, 36(6), 720–735. https://doi.org/10.3102/1076998610396890
}
