% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robust.R
\name{item.prob}
\alias{item.prob}
\title{Item Response Probability
Computes item response probabilities for select IRT models (1PL, Rasch, 2PL, MIRT, GRM, and MGRM), given ability and item parameters.
by constructing the appropriate linear predictors and applying the logistic function. Returns item response probabilities for dichotomous data or item category response probabilities for polytomous data.}
\usage{
item.prob(theta, model, ipars, D = 1.7)
}
\arguments{
\item{theta}{A numeric vector or matrix of latent trait values.}

\item{model}{A character string specifying which IRT model to use.
* 'Rasch' - Allows for item difficulty parameters to vary across items (Rasch, 1960).
* '1PL' - 1-parameter logistic model. Allows for an item discrimination parameter that is constrained to be equal for all items and item difficulty parameters that vary across items. The Rasch model is a special case of the 1PL model in which all item discrimination values are constrained to 1.
* '2PL' - 2-parameter logistic model. Allows for both item discrimination and difficulty parameters to vary across items (Birnbaum, 1968).
* 'MIRT' - Multidimensional extension of the 2PL model, where item slopes can vary across multiple latent dimensions and items, and difficulty parameters which vary only with the item (McKinley & Reckase, 1983; Muraki and Englelhard, 1985).
* 'GRM' - Graded response model for ordered polytomous items, allowing item discrimination parameters to vary across items and item slopes to vary across items and category thresholds (Samejima, 1969).
* 'MGRM' - Multidimensional graded response model, extending Samejima’s GRM to multiple latent dimensions. Allows item discrimination parameters to vary across items and latent dimensions and item slopes to vary across items and category thresholds. (Muraki & Carlson, 1995).}

\item{ipars}{A matrix of item parameters. See examples for how to structure the columns of the matrix based on the model utilitized.}

\item{D}{A positive scaling constant used for scaling the normal ogive model. Defaults to 1.7; alternatively is often set to 1.0.}
}
\value{
For model accommodating dichotomous data ("1PL", "2PL", "MIRT"), returns an \eqn{N \times J} matrix of response probabilities P(X = 1)

For models accommodating polytomous data ("GRM", "MGRM"), returns a list with: {pstar}: an array of cumulative probabilities \eqn{P^*(X \geq k)} 
        and {P}: an array of category probabilities \eqn{P(X = k)}
}
\description{
Item Response Probability
Computes item response probabilities for select IRT models (1PL, Rasch, 2PL, MIRT, GRM, and MGRM), given ability and item parameters.
by constructing the appropriate linear predictors and applying the logistic function. Returns item response probabilities for dichotomous data or item category response probabilities for polytomous data.
}
\examples{
# Rasch case
N <- 3
J <- 5
theta <- rnorm(N) # generate ability values
ipars <- rnorm(J)  # generate item difficulties
item.prob(theta, "Rasch", ipars)
# 1PL case
N <- 4 # subjects
J <- 6 # items
theta <- rnorm(N) # generate ability values
ipars <- cbind(a = rep(1.2, J), # set item discrimination
               b = rnorm(J)) # generate item difficulties
item.prob(theta, "1PL", ipars)
# 2PL case
N <- 3 # subjects
J <- 5 # items
theta <- rnorm(N) # generate ability values
ipars <- cbind(a = runif(J, 0.5, 2), # generate item discrimination
               b = rnorm(J)) # generate item difficulty
item.prob(theta, "2PL", ipars) 
# MIRT case
N <- 2 # subjects
J <- 7 # items
L <- 3 # dimensions
theta <- matrix(rnorm(N * L), ncol = L) # N x L ability matrix
ipars <- cbind(matrix(runif(J * L, 0, 1), ncol = L), d = rnorm(J)) # generate slopes + intercept
item.prob(theta, "MIRT", ipars)
# GRM case
N <- 3 # subjects
J <- 5 # items
K <- 4 # categories
theta <- rnorm(N)
a <- runif(J, 0.5, 2) # J discriminations
b_raw <- matrix(rnorm(J * (K - 1)), nrow = J)  
b <- t(apply(b_raw, 1, sort)) # Sort thresholds within each item (row-wise)
ipars <- cbind(a, b) # Combine into a J x K matrix
item.prob(theta, "GRM", ipars)
# MGRM case
N <- 3 # subjects
J <- 5 # items
L <- 2 # dimensions
K <- 4 # categories
theta <- matrix(rnorm(N * L), ncol = L) 
a <- matrix(runif(J * L, 0, 1), ncol = L) # slopes 
b_raw <- matrix(rnorm(J * (K - 1)), nrow = J) # thresholds
b <- t(apply(b_raw, 1, sort))
ipars <- cbind(a, b)
item.prob(theta, "MGRM", ipars)
}
\references{
Birnbaum, A. (1968). Some latent trait models and their use in inferring an examinee’s ability. In F. M. Lord & M. R. Novick (Eds.), Statistical Theories of Mental Test Scores (pp. 397–479). \emph{Addison‑Wesley}.

Lord, F. M., & Novick, M. R. (1968). Statistical theories of mental test scores. \emph{Addison-Wesley}.

Lord, F. M. (1980). Applications of item response theory to practical testing problems. \emph{Erlbaum}. https://doi.org/10.4324/9780203056615.

McKinley, R. L., & Reckase, M. D. (1983). An extension of the two‑parameter logistic model to the multidimensional latent space. \emph{Psychometrika}, 48(3), 369–382.

Mosteller, F., & Tukey, J. W. (1977). Data Analysis and Regression: A Second Course in Statistics. \emph{Addison‑Wesley}.

Muraki, E., & Engelhard, G. (1985). Full-information item factor analysis: Applications of EAP scores. \emph{Applied Psychological Measurement}, 9(4), 417–430

Rasch, G. (1960). Probabilistic models for some intelligence and attainment tests. \emph{Danish Institute for Educational Research}, 184.

Samejima, F. (1969). Estimation of latent ability using a response pattern of graded scores. \emph{Psychometrika Monograph Supplement, 34} (4, Pt. 2), 100–100.
}
