% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robust.R
\name{msr}
\alias{msr}
\title{Deprecated - see residual. Modified Standardized Residual}
\usage{
msr(dat, theta, a, b)
}
\arguments{
\item{dat}{\eqn{N \times J} matrix of data, beginning indexing at 1 for \eqn{N} respondents and \eqn{J} items}

\item{theta}{\eqn{N \times L} matrix of \eqn{\theta}s, for \eqn{L} dimensions}

\item{a}{\eqn{J \times L} matrix of discrimination parameters}

\item{b}{\eqn{J \times (K-1)} matrix of category threshold parameters, for \eqn{K} Likert-type categories}
}
\value{
\eqn{N \times J} matrix of residuals
}
\description{
Calculate the modified standardized residual (MSR) used to quantify misfit between a response and the IRT model (Yu & Cheng, 2019): 
\deqn{ \frac{y_j - E(Y_j | \hat{\theta})}{P(y_j|\hat{\theta})}}
given data and parameters that follow the multidimensional graded response model (MGRM; Muraki & Carlson, 1995). Handles unidimensional and dichotomous data as well.
}
\references{
Yu, X & Cheng, Y. A change-point analysis procedure based on weighted residuals to detect back random responding. \emph{Psychological Methods} (Oct. 2019), pp. 658–674. DOI: 10.1037/met0000212.

Muraki, E. & Carlson, J. E. Full-Information Factor Analysis for Polytomous Item Responses. \emph{Applied Psychological Measurement} (Mar. 1995), pp. 73–90. DOI: 10.1177/014662169501900109.
}
