% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{MTurkHeight}
\alias{MTurkHeight}
\title{MTurk Height Data}
\format{
A data frame with 1402 rows (participants) and 34 columns (survey items and follow-up information). 
\itemize{
  \item{\code{country}}{ The country the user's network connection was based.}
  \item{\code{engnat}}{ Whether the respondent reported English as their native language (0 = no, 1 = yes).}
  \item{\code{age}}{ Age in years.}
  \item{\code{gender}}{ Gender indicator.)}
  \item{\code{Q1}}{ I am tall.}
  \item{\code{Q2}}{ I am short.}
  \item{\code{Q3}}{ I have to stand on a stool to reach tall kitchen shelves.}
  \item{\code{Q4}}{ I have to stand in the back in group photos to not cover up other people.}
  \item{\code{Q5}}{ I hit my head on low ceilings.}
  \item{\code{Q6}}{ I rarely meet people with more height than me.}
  \item{\code{Q7}}{ I'm kind of a midget.}
  \item{\code{Q8}}{ Airplane seats never have enough room for my long legs.}
  \item{\code{Q9}}{ When I hug people, my head is underneath their chin.}
  \item{\code{Q10}}{ I have gangly limbs.}
  \item{\code{Q11}}{ I have been sent to the hospital by an electric shock.}
  \item{\code{Q12}}{ I own a goat.}
  \item{\code{Q13}}{ I know the 'happy birthday to you..' song.}
  \item{\code{Q14}}{ I have been asked for money by beggars.}
  \item{\code{Q15}}{ I prefer to play it safe and avoid danger.}
  \item{\code{Q16}}{ I prefer variety to routine.}
  \item{\code{Q17}}{ I rarely clean house.}
  \item{\code{Q18}}{ I rarely complain.}
  \item{\code{Q19}}{ I rarely overindulge.}
  \item{\code{Q20}}{ I accept what others say.}
  \item{\code{Q21}}{ I enjoy being part of a loud crowd.}
  \item{\code{Q22}}{ I offend no one.}
  \item{\code{Q23}}{ I see that nobody gets left out.}
  \item{\code{Q24}}{ I try not to deceive others.}
  \item{\code{Q25}}{ I try out new things.}
  \item{\code{Q26}}{ I will push people around to get what I want.}
  \item{\code{feet}}{ Height reported in feet (if the respondent used imperial units).}
  \item{\code{inch}}{ Height reported in inches (if the respondent used imperial units).}
  \item{\code{cm}}{ Height reported in centimeters (if the respondent used metric units).}
  \item{\code{submittime}}{ The time (PST) the survey was submitted.}
}
}
\source{
Open Psychometrics (2019, December 29). A quality comparison of data collected on this website to data collected on Amazon Mechanical Turk. https://openpsychometrics.org/_rawdata/validity/
}
\usage{
data(MTurkHeight)
}
\description{
A dataset containing 26 survey items collected from Amazon Mechanical Turk (AMT). Additional demographic variables are also included.
Respondents were recruited from MTurk and were paid $0.20 to participate. To qualify, workers were required to have completed at least 100 prior human intelligence tasks (HITs)  
with an acceptance rate of 97% or higher.
}
\details{
The item data followed a 5-point Likert scale, with the following labels: 1 = Strongly disagree, 2 = Disagree, 3 = Neither agree nor disagree, 4 = Agree, and 5 = Strongly agree. A 0 indicates no response.
The first four items include demographic information. Q1 through Q10 measure height, while Q13 through Q26 measure general personality characteristics. 
The last four items ask users to report their hight and contain a record of submission time. 
Q11 and Q12 are bogus items, designed to detect inattentive responders, or those who endorse the item although it is unlikely to be true.
Due to factors such as disengagement and low motivation that are common on low stakes assessment, MTurk data may be especially susceptible to aberrant response behavior. 
When fitting an item response theory model, robust estimation may be applied to reduce bias in the latent trait estimate for respondents exhibiting aberrant responses.
However, when a subject's response pattern contains aberrant responses for the majority of or all items (e.g., "bot" responding), robust estimation will not be effective in mitigating the aberrances and unable to reduce the bias in the latent trait.
}
\keyword{datasets}
