% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robust.R
\name{probs.calc.grm}
\alias{probs.calc.grm}
\title{GRM Response Probability \eqn{P^*} and Response Category Probabilities (P) Calculation (deprecated - see item.prob)}
\usage{
probs.calc.grm(thetas, a, b)
}
\arguments{
\item{thetas}{A vector of latent traits for \eqn{N} subjects}

\item{a}{A vector of discrimination parameters for \eqn{J} test items}

\item{b}{A matrix of category threshold parameters, with the number of rows corresponding to \eqn{J} test items and the number of columns corresponding to \eqn{K-1} thresholds between \eqn{K} categories}
}
\value{
pstar A \eqn{J \times (K-1) \times N} array of category threshold probabilities for \eqn{J} items, \eqn{K} categories, and \eqn{N} subjects

P a \eqn{J \times K \times N} array of probabilities of responding in each of \eqn{K} categories for \eqn{J} items, across \eqn{N} subjects
}
\description{
Generate \eqn{P^*}, the threshold probability, and \eqn{P}, the probability of responding in each category for a vector of latent traits
using parameters needed for the GRM (Samejima, 1969).
The probability that a subject responds in or above a category \eqn{k} for item \eqn{j} is \eqn{P^*_{jk}(\theta) = \frac{1}{1+ e^{-a_j (\theta-b_{jk})}}},
for \eqn{K} categories and \eqn{K-1} threshold parameters  (\eqn{b_{j1}, ... b_{j,K-1}}), where \eqn{b_{jk}} separates response category \eqn{k} and \eqn{k+1} (\eqn{k=1,..K-1}) (Embretson & Reise, 2000).
\eqn{a_j} is the item discrimination parameter.  The probability of endorsing exactly category \eqn{k} is \eqn{P_{jk}(\theta) = P^*_{j,k}(\theta) - P^*_{j,k+1}(\theta),} where \eqn{P^*_{j1}(\theta) \equiv 1.0} and \eqn{P^*_{jK}(\theta) \equiv 0.0.}
}
\examples{
thetas <- rnorm(15)  # Example latent traits for 15 subjects
a <- runif(10, 0.5, 1.5)  # Example discrimination parameters for 10 items
b <- t(apply(matrix(runif(10*4, -2.5,2.5), nrow = 10, ncol = 4), 1, sort))  # Example threshold parameters for 10 items and 4 thresholds (5 categories)
result <- probs.calc.grm(thetas, a, b)  # Calculate probabilities
}
\references{
Embretson, S. E., & Reise, S. P. (2000). \emph{Item response theory for psychologists.} Mahwah, N.J: L. Erlbaum Associates.

Samejima, F. (1969). Estimation of latent ability using a response pattern of graded scores. \emph{Psychometrika Monograph Supplement, 34} (4, Pt. 2), 100–100.
}
