% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robust.R
\name{probs.calc}
\alias{probs.calc}
\title{Response Probability Calculation (MIRT) -- deprecated (see item.prob)}
\usage{
probs.calc(thetas, a, d)
}
\arguments{
\item{thetas}{A vector of length \emph{L} containing a subject's latent traits, where \emph{L} is the number of test dimensions}

\item{a}{A \eqn{J \times L} matrix containing fixed item slope parameters for \emph{L} dimensions and \emph{J} test items}

\item{d}{A vector of length \emph{J} containing fixed intercept parameters, for \emph{J} test items}
}
\value{
P A vector of probabilities for endorsing each of \emph{J} test items

residual A vector of residuals detecting misfit between the subject and \emph{J} test items, according to the formula \eqn{r_{ij} = \textbf a_j\boldsymbol{\theta}_i + d_j} (see theta.est)
}
\description{
Calculate the item response probabilities given the item parameters and latent trait vector for person \emph{i}, which is
\eqn{\boldsymbol{\theta}_i = ({\theta}_{i1}, {\theta}_{i2}, ... {\theta}_{iL})'} for \emph{L} dimensions, according to the multidimensional IRT model (MIRT; McKinley & Reckase, 1983):
\deqn{P (X_{ij}=1 | \boldsymbol{\theta}_i) =  \frac{1}{1+ e^{-1.7(\boldsymbol{a_j}\boldsymbol\theta+d_j)}}} for \eqn{j=1,...,J} items.
An item \emph{j} has slope parameters \eqn{\boldsymbol{a}_j=(a_{1j}, a_{2j}, ..., a_{Lj})'} and intercept parameter \eqn{d_j}.
}
\references{
McKinley, R. L., & Reckase, M. D. (1983, August). \emph{An Extension of the Two-Parameter Logistic Model to the Multidimensional Latent Space} (Research No. ONR83-2). Iowa City, IA: American College Testing Program.
}
