% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robust.R
\name{probs.gen.grm}
\alias{probs.gen.grm}
\title{Response Probability Calculation (GRM) and response category probabilities (P)}
\usage{
probs.gen.grm(thetas, a, b)
}
\arguments{
\item{thetas}{A vector of latent traits for \eqn{m} subjects}

\item{a}{A vector of discrimination parameters for \eqn{n} test items}

\item{b}{A matrix of category threshold parameters, with the number of rows corresponding to \eqn{n} test items and the number of columns corresponding to \eqn{K-1} thresholds between \eqn{K} categories}
}
\value{
pstar A \eqn{n \times (K-1) \times m} array of category threshold probabilities for \eqn{n} items, \eqn{K} categories, and \eqn{m} subjects

P a \eqn{n \times K \times m} array of probabilities of responding in each of \eqn{K} categories for \eqn{n} items, across \eqn{m} subjects
}
\description{
This function generates \eqn{P^*}, the threshold probability, and \eqn{P}, the probability of responding in each category for a vector of latent traits
}
