% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robust.R
\name{probs.gen}
\alias{probs.gen}
\title{Response Probability Calculation (2PL)}
\usage{
probs.gen(thetas, a, d)
}
\arguments{
\item{thetas}{A vector of length \emph{p} containing students’ latent traits, where \emph{p} is the number of test dimensions}

\item{a}{A \eqn{n \times p} matrix containing fixed item slope parameters for \emph{p} dimensions and \emph{n} test items}

\item{d}{A vector of length \emph{n} containing fixed intercept parameters, for \emph{n} test items}
}
\value{
P A vector of probabilities for endorsing each of \emph{n} test items

residual A vector of residuals detecting misfit between the subject and \emph{n} test items, according to the formula \eqn{r_i = \textbf a_i\boldsymbol\theta + d_i} (see theta.est)
}
\description{
This function calculates the probabilities of getting items correct given the item parameters and a person's abilities
}
