% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robust.R
\name{probs.mgrm}
\alias{probs.mgrm}
\title{Item category response probability for MGRM (deprecated - see item.prob)}
\usage{
probs.mgrm(thetas, a, b, D = 1.7)
}
\description{
Item category response probability for MGRM (deprecated - see item.prob)
}
