% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robust.R
\name{robust.item.plots}
\alias{robust.item.plots}
\title{Comparing Robust with MML Item Parameter Estimates}
\usage{
robust.item.plots(dat, ...)
}
\arguments{
\item{dat}{A \eqn{N\times J} matrix of response data}

\item{...}{Additional arguments to be passed to \code{mirt()}}
}
\value{
A list containing one plot for each item parameter, structured as \code{`Parameter Estimates for a1`}, \code{`Parameter Estimates for d`}, etc., in addition to
\describe{
  \item{\code{`MML Item Parameter Estimates`}}{ Matrix of item parameter estimates obtained with MML as defined by \code{mirt()}, each row corresponding to an item.}
  \item{\code{`RMML Item Parameter Estimates`}}{ Matrix of item parameter estimates obtained with robust MML as defined by \code{mirt()}, each row corresponding to an item.}
}
}
\description{
The following function returns a plot comparing the robust estimate (Hong & Cheng, 2019) of an item parameter against the corresponding marginal maximum likelihood (MML) estimate for all items.
The item parameters are labeled according to their label in \code{mirt()}.
}
\details{
When the model fits the data, item parameters estimated with RMML should be approximately equivalent to those estimated with MML, thereby roughly following the identity line. 
         When the two estimates differ greatly for a parameter on an item, aberrant responding may be present such that its effects are mitigated by robust weighting. 
         The plots returned by \code{robust.item.plots()} can therefore be used to spot items that are greatly affected by robust estimation, for further investigation.
}
\examples{

# SAT12 Data: 3-Dimensional MIRT Model
SAT12[SAT12 == 8] <- NA #set 8 as a missing value
# Correct answer key
library(mirt)
key <- c(1,4,5,2,3,1,2,1,3,1,2,4,2,1,5,3,4,4,1,4,3,3,4,1,3,5,1,3,1,5,4,5) 
scoredSAT12 <- key2binary(SAT12, key)
specific <- c(2, 3, 2, 3, 3, 2, 1, 2, 1, 1, 1, 3, 1, 3, 1, 2, 1, 1, 3, 3, 1, 1, 3, 1, 3, 3, 1, 3, 2, 3, 1,2) #which factor each item loads on

plots_out <- robust.item.plots(scoredSAT12, model=specific)

# View three discrimination parameters and intercept
plots_out$`Parameter Estimates for a1`
plots_out$`Parameter Estimates for a2`
plots_out$`Parameter Estimates for a3`
plots_out$`Parameter Estimates for d`

}
\references{
Hong, M., & Cheng, Y. (2019). Robust maximum marginal likelihood (RMML) estimation for item response theory models. Behavior Research Methods, 51(2), 573–588. https://doi.org/10.3758/s13428-018-1150-4
}
