% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robust.R
\name{robust.rt}
\alias{robust.rt}
\title{Robust Estimation of Working Speed}
\usage{
robust.rt(
  dat,
  ipars,
  weight.type = "equal",
  tuning.par = NULL,
  custom.weights = NULL,
  iter = 30,
  tol = 1e-04
)
}
\arguments{
\item{dat}{A \eqn{N\times J} matrix of numerical response time data}

\item{ipars}{Item-level parameters for the log-normal model, structured with the first column containing discrimination parameters \eqn{\alpha_j} and the second column containing time-intensity parameters \eqn{\Beta_j} for \eqn{j=1,...,J}.}

\item{weight.type}{Weighting scheme: \code{"equal"} (default), \code{"Huber"},
\code{"bisquare"}, or \code{"custom"}.}

\item{tuning.par}{Tuning parameter for Huber or bisquare weights.  Required when
\code{weight.type} is not \code{"equal"} or \code{"custom"}.}

\item{custom.weights}{An \eqn{N \times J} numeric matrix of user-specified weights
in \eqn{[0, 1]}. Required when \code{weight.type = "custom"}.}

\item{iter}{Maximum number of iterations. Default is 30.}

\item{tol}{Convergence tolerance on the absolute change in working speed estimate between two iterations.
Default is 0.0001.}
}
\value{
A list containing
\describe{
  \item{\code{tau}}{ (\eqn{N \times 1}) matrix of working speed estimates.}
  \item{\code{residuals}}{ (\eqn{N \times J}) matrix of item-level residuals (\eqn{r_j}) capturing the difference between the observed and model-expected response time for each individual.}
  \item{\code{nonconvergence}}{ (\eqn{N \times 1}) Nonconvergence indicator (0 = converged, 1 = not converged).}
}
}
\description{
To reduce the influence of spuriously slow or spuriously fast response times on the estimate of an examinee's working speed (\eqn{\tau}), robust M-estimates can be obtained (Hong et al., 2020).
Under the log-normal model (van der Linden, 2006), the log response time (\eqn{t_j}) for an examinee on item \eqn{j} (\eqn{j=1,2,...J}), follows 
\eqn{\log(t_j) | \tau ~ N(\Beta_j - \tau, \frac{1}{\alpha_j^2})},
assuming independence of log item response times for the examinee. The probability density function for the model is then given by
\eqn{f(t_i|\tau) = \frac{\alpha_j}{t_j \sqrt{2\pi}} \exp{(-\frac{1}{2}(a_j(\log t_j - (\Beta_j - \tau)))^2)}}
The robust maximum likelihood estimate of \eqn{\tau} is 
\eqn{\hat{\tau}^{RML} = \frac{\sum_{j=1}^J w(r_j) \alpha_j^2 (\Beta_j - \log t_j)}{\sum_{j=1}^J w(r_j) \alpha_j^2 }}
where weights \eqn{w(\cdot)} are defined by either the Huber (Huber, 1981) or bisquare (Mosteller & Tukey, 1977) weight functions (see \code{huber()} and \code{bisquare()} functions for more detail).
A residual capturing the difference between the observed and model-expected response time is calculated with \eqn{r_j = \alpha_j(\log(t_{ij}) - \Beta_j - \tau_i)}.
Although \eqn{\hat{\tau}^{RML}} yields a closed-form solution, the residuals and corresponding weights are updated based on the previously estimated \eqn{\hat{\tau}^{RML}} in an iterative manner, with the initial value set at the maximum likelihood estimate (where all weights are fixed at 1).
Convergence occurs when the absolute change in \eqn{\hat{\tau}^{RML}} between two iterations is within the specified tolerance.
}
\details{
The working speed parameter, \eqn{tau}, captures how "speedily" a subject responds. 
         In effect, a larger \eqn{tau} corresponds to quicker responses, while lower \eqn{tau}s correspond to slower responses.
         When an anomalous handful of response times are spuriously quicker than others, the MLE of \eqn{tau} may become inflated (suggesting more speediness), and the robust estimate counteracts this bias, yielding a smaller estimate.
         Conversely, spuriously slower responses may yield a smaller MLE than the true working speed, while the robust estimate works to increase the MLE closer to the truth.
         For identifiability purposes, \eqn{tau} is assumed to have a mean of 0.
}
\examples{

# Example with Canadian PISA data
library(pisaRT)
dat<- pisaW[,26:37]
# Estimate item parameters for the log-normal model under normal responding
library(LNIRT)
mod.ln <- LNRT(RT = dat, data = dat)$Post.Means
ipars<-cbind(mod.ln$Time.Discrimination, mod.ln$Time.Intensity)
# Reduce response time by 1/4 for people as if they had item preknowledge on items 2, 4, 6, 8
dat[,c(2, 4, 6, 8)]<-dat[,c(2, 4, 6, 8)] - log(4)

# Robust estimation of tau with aberrant data
out.rt.ab<-robust.rt(dat = exp(dat), ipars, weight.type = "bisquare", tuning.par = 4)

# Robust estimation of tau with normal response data
out.rt<-robust.rt(dat = exp(dat), ipars, weight.type = "equal")
head(cbind(out.rt.ab$tau, out.rt$tau))

# Robust estimates are lower, while MLEs suggest speedier responses

}
\references{
Huber, P. (1981). \emph{Robust Statistics}. John Wiley & Sons, Inc.

Mosteller, F., & Tukey, J. W. (1977). \emph{Data Analysis and Regression: A Second Course in Statistics}.
  Addison-Wesley Publishing Company.
}
