% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{simBRR2PL}
\alias{simBRR2PL}
\title{Simulated 2PL data with back random responses}
\format{
A list containing:
\describe{
  \item{\code{dat}}{ (\eqn{150 \times 30}) matrix of simulated dichotomous data.}
  \item{\code{ipars}}{ (\eqn{30 \times 2}) data frame of true item parameters for the 2PL model: one column for item discrimination and one column for item difficulty.}
  \item{\code{ab.indices}}{ Vector of row numbers corresponding to the designated aberrant examinees.}
 }
}
\usage{
data(simBRR2PL)
}
\description{
Data simulated from a 2PL model for 1500 subjects on 30 items. For 20% of examinees on the last 30% of items, correct responses were generated with a 20% success rate.
}
\examples{

## Not run: 


}
\references{
Clark, M. E., Gironda, R. J., & Young, R. W. (2003). Detection of back random responding: Effectiveness of MMPI-2 and Personality Assessment Inventory validity indices. Psychological Assessment, 15(2), 223–234. https://doi.org/10.1037/1040-3590.15.2.223

Yu, X., & Cheng, Y. (2019). A change-point analysis procedure based on weighted residuals to detect back random responding. Psychological Methods, 24(5), 658–674. https://doi.org/10.1037/met0000212
}
\keyword{datasets}
