% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{simBRRMGRM}
\alias{simBRRMGRM}
\title{Simulated MGRM data with back random responding}
\format{
A list containing:
\describe{
  \item{\code{dat}}{ (\eqn{2000 \times 50}) matrix of simulated Likert-type data.}
  \item{\code{ipars}}{ (\eqn{50 \times 7}) data frame of true item parameters for the GRM: three columns for item discrimination and four columns for the four category thresholds.}
  \item{\code{ab.indices}}{ Vector of row numbers corresponding to the designated aberrant examinees.}
 }
}
\usage{
data(simBRRMGRM)
}
\description{
Data simulated from a MGRM model for 2000 subjects on 50 items. Likert-type responses are scored 1 to 5, and the model follows a 3-dimensional simple structure in which each item loads onto only one dimension. 
For 40% of examinees, their responses to the last 40% of items were generated as if they were randomly responding, with a uniform probability of selecting any of the five categories.
}
\examples{

## Not run: 


}
\references{
Clark, M. E., Gironda, R. J., & Young, R. W. (2003). Detection of back random responding: Effectiveness of MMPI-2 and Personality Assessment Inventory validity indices. Psychological Assessment, 15(2), 223–234. https://doi.org/10.1037/1040-3590.15.2.223

Yu, X., & Cheng, Y. (2019). A change-point analysis procedure based on weighted residuals to detect back random responding. Psychological Methods, 24(5), 658–674. https://doi.org/10.1037/met0000212
}
\keyword{datasets}
