% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{simCheating2PL}
\alias{simCheating2PL}
\title{Simulated 2PL data with cheating / spuriously high responses}
\format{
A list containing:
\describe{
  \item{\code{dat}}{ (\eqn{1000 \times 30}) matrix of simulated dichotomous data.}
  \item{\code{ipars}}{ (\eqn{30 \times 2}) data frame of true item parameters for the 2PL model: one column for item discrimination and one column for item difficulty.}
  \item{\code{ab.indices}}{ Vector of row numbers corresponding to the aberrant examinees.}
 }
}
\usage{
data(simCheating2PL)
}
\description{
Data simulated from a 2PL model for 1000 subjects on 30 items. For the 20% of examinees with the lowest true \eqn{\theta}, they received correct responses on the most difficult 10% of items.
}
\examples{

## Not run: 


}
\references{
Wang, C., Xu, G., Shang, Z., & Kuncel, N. (2018). Detecting Aberrant Behavior and Item Preknowledge: A Comparison of Mixture Modeling Method and Residual Method. Journal of Educational and Behavioral Statistics, 43(4), 469–501. https://doi.org/10.3102/1076998618767123
}
\keyword{datasets}
