% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{simCheatingMIRT}
\alias{simCheatingMIRT}
\title{Simulated MIRT data with cheating}
\format{
A list containing:
\describe{
  \item{\code{dat}}{ (\eqn{2000 \times 50}) matrix of simulated Likert-type data.}
  \item{\code{ipars}}{ (\eqn{50 \times 3}) data frame of true item parameters for the GRM: two columns for item discrimination and one column for item threshold.}
  \item{\code{ab.indices}}{ Vector of row numbers corresponding to the designated aberrant examinees.}
 }
}
\usage{
data(simCheatingMIRT)
}
\description{
Data simulated from a MGRM model for 2000 subjects on 50 items. The model is 2-dimensional, following a complex structure in which each item is allowed to load onto each dimension. For the 20% of examinees with the lowest \eqn{\theta}, their responses to the most difficult 20% of items were coded as correct, as if they had cheated to obtain the correct answer.
}
\examples{

## Not run: 


}
\references{
Meijer, R. R. (1996). Person-Fit Research: An Introduction. Applied Measurement in Education, 9(1), 3–8. https://doi.org/10.1207/s15324818ame0901
}
\keyword{datasets}
