% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{simReverseGRM}
\alias{simReverseGRM}
\title{Simulated GRM data with improper reverse coding}
\format{
A list containing:
\describe{
  \item{\code{dat}}{ (\eqn{1000 \times 30}) matrix of simulated Likert-type data.}
  \item{\code{ipars}}{ (\eqn{30 \times 5}) data frame of true item parameters for the GRM: one column for item discrimination and four columns for the four category thresholds.}
  \item{\code{ab.indices}}{ Vector of row numbers corresponding to the designated aberrant examinees.}
 }
}
\usage{
data(simReverseGRM)
}
\description{
Data simulated from a GRM model for 1000 subjects on 30 items. Likert-type responses are scored 1 to 5. For 20% of examinees, their responses to 30% of items were reverse coded as if they were inatentive to the reverse-wording of the item.
}
\examples{

## Not run: 


}
\references{
Hughes, G. D. (2009). The Impact of Incorrect Responses to Reverse-Coded Survey Items. Research in the Schools, 14
}
\keyword{datasets}
