% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{simSuboptimal2PL}
\alias{simSuboptimal2PL}
\title{Simulated 2PL data with suboptimal responses}
\format{
A list containing:
\describe{
  \item{\code{dat}}{ (\eqn{500 \times 39}) matrix of simulated dichotomous data.}
  \item{\code{ipars}}{ (\eqn{30 \times 2}) data frame of true item parameters for the 2PL model: one column for item discrimination and one column for item difficulty.}
  \item{\code{ab.indices}}{ Vector of row numbers corresponding to the designated aberrant examinees.}
 }
}
\usage{
data(simSuboptimal2PL)
}
\description{
Data simulated from a 2PL model for 500 subjects on 30 items. For 40% of examinees, data was generated as if the subject responded under suboptimal ability. 
For these individuals, data was generated according to \eqn{\theta_{new} = \theta_{true} - 1}, where \eqn{\theta_{true}} is the true latent trait.
}
\examples{

## Not run: 


}
\references{
C. Schuster and K.-H. Yuan. Robust Estimation of Latent Ability in Item Response Models. Journal of Educational and Behavioral Statistics, 36(6):720–735, Dec. 2011. ISSN 1076-9986, 1935-1054. doi: 10.3102/1076998610396890.
}
\keyword{datasets}
