% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{simWarmUp2PL}
\alias{simWarmUp2PL}
\title{Simulated 2PL data with warm-up / sleeping behavior}
\format{
A list containing:
\describe{
  \item{\code{dat}}{ (\eqn{1000 \times 30}) matrix of simulated dichotomous data.}
  \item{\code{ipars}}{ (\eqn{30 \times 2}) data frame of true item parameters for the 2PL model: one column for item discrimination and one column for item difficulty.}
  \item{\code{ab.indices}}{ Vector of row numbers corresponding to the designated aberrant examinees.}
 }
}
\usage{
data(simWarmUp2PL)
}
\description{
Data simulated from a 2PL model for 1000 subjects on 30 items. For 30% of examinees, they received an incorrect response on the first 30% of items as if they were "warming" or "waking" up to the test content or format.
}
\examples{

## Not run: 


}
\references{
Meijer, R. R. (2002). Outlier Detection in High-Stakes Certification Testing. Journal of Educational Measurement, 39(3), 219–233. https://doi.org/10.1111/j.1745-3984.2002.tb01175.x
}
\keyword{datasets}
