% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robust.R
\name{std.err.dichotomous}
\alias{std.err.dichotomous}
\title{Deprecated - see standard.erors. Calculate standard errors of ability estimates for MIRT data}
\usage{
std.err.dichotomous(
  theta,
  d,
  a,
  dat,
  D = 1.7,
  residual = "standardized",
  weight.type = "equal",
  tuning.par = NULL
)
}
\arguments{
\item{theta}{Vector of latent traits (abilities) for an individual}

\item{d}{Vector of difficulty parameters for the items}

\item{a}{Matrix of discrimination parameters for the items (rows are items, columns are dimensions)}

\item{D}{A positive scaling constant used for scaling the normal ogive model. Defaults to 1.7; alternatively is often set to 1.0.}
}
\value{
Vector of standard errors
}
\description{
Calculate the standard errors of ability estimates using the Fisher Information matrix for multidimensional dichotomous data using the MIRT model
}
\examples{
#Define the parameters
theta <- c(0.5, -1.2) # Example vector of latent traits
d <- c(0.0, -0.5, 1.0) # Example vector of difficulty parameters for 3 items
a <- matrix(c(1.0, 0.5,
              0.7, 1.2,
              1.5, 1.0), nrow = 3, byrow = TRUE) # Example matrix of discrimination parameters for 3 items
dat <- c(1,1,0)
D <- 1.7 # Scaling constant
#Calculate the standard errors
std.errs <- std.err.dichotomous(theta, d, a, D)
print(std.errs)
}
