% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robust.R
\name{std.err.mgrm}
\alias{std.err.mgrm}
\title{Standard error for MGRM ability estimation}
\usage{
std.err.mgrm(
  theta,
  d,
  a,
  dat,
  D = 1.7,
  weight.type = "equal",
  tuning.par = NULL
)
}
\description{
Standard error for MGRM ability estimation
}
