% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robust.R
\name{std.err.poly}
\alias{std.err.poly}
\title{Deprecated - see standard.erors. Calculate standard errors of ability estimates for GRM data}
\usage{
std.err.poly(theta, dat, b, a, weight.type, tuning.par, D = 1.7)
}
\arguments{
\item{theta}{Vector of latent traits (abilities) for an individual}

\item{dat}{Vector of observed item responses for the individual, of length $J$ for $J$ items}

\item{b}{Matrix of threshold parameters for the items, with number of rows corresponding to the number of items and number of columns corresponding to the number of thresholds}

\item{a}{Vector of discrimination parameters for the items of length $J$}

\item{weight.type}{Type of weighting function to be used: "equal", "Huber", or "bisquare"}

\item{tuning.par}{Tuning parameter to be used for the Huber or bisquare weight function}

\item{D}{A positive scaling constant used for scaling the normal ogive model. Defaults to 1.7; alternatively is often set to 1.0.}
}
\value{
Standard Error Standard error of $\theta$ based on the Fisher information (expected information) matrix

Sandwich SE Huber-White sandwich estimator of the standard error
}
\description{
Calculate the standard errors of ability estimates using the Fisher Information matrix or Huber-White sandwich standard errors for Likert-type response data using the GRM model
}
