% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robust.R
\name{theta.recov}
\alias{theta.recov}
\title{Ability Estimation Function using MLE}
\usage{
theta.recov(
  dat,
  a,
  d,
  iter = 100,
  theta.initial = rep(0, ncol(a)),
  weight.category = "normal",
  tuning.par = NULL
)
}
\arguments{
\item{dat}{Response data to load}

\item{a}{Matrix of slope parameters}

\item{d}{Array of intercept parameters}

\item{iter}{Max number of iterations. Default is 100}

\item{theta.initial}{Array of initial thetas. Default is 0s}

\item{weight.category}{The weighting strategy to use, "normal", "bisquare" and "Huber". Default is "normal", which is equally weighted.}

\item{tuning.par}{The tuning parameter for "bisquare" or "Huber" weighting functions}
}
\value{
Array of estimated person abilities
}
\description{
This function return the list of ability estimations based on the given weighting function
}
\details{
Huber weighting function
}
