% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robust.R
\name{theta_plots}
\alias{theta_plots}
\title{Plot to compare robust estimates with MLE}
\usage{
theta_plots(
  dat,
  a,
  d = NULL,
  b = NULL,
  iter = 30,
  cutoff = 0.01,
  H = NULL,
  B = NULL,
  same.plot.dim = F,
  same.plot = T,
  type
)
}
\arguments{
\item{dat}{\eqn{J \times N} matrix of response data for \emph{J} items and \emph{N} subjects}

\item{a}{\eqn{J \times L} matrix of slope parameters for \emph{J} items and \emph{L} dimensions (\emph{L=1} if using the GRM or unidimensional 2PL model)}

\item{d}{If type = “MIRT”, a vector of discrimination parameters for \emph{J} items}

\item{b}{If type = “GRM”, an \eqn{J \times (K-1)} matrix of intercept parameters}

\item{iter}{Maximum number of iterations. Default is 30}

\item{cutoff}{Threshold value to terminate the iteration when the likelihood changes below this value, which means that the estimation is converged. Default is 0.01.}

\item{H}{Huber tuning parameter}

\item{B}{Bisquare tuning parameter}

\item{same.plot.dim}{If TRUE and type = “MIRT”, estimates across all \emph{L} dimensions will be plotted on the same graph. If FALSE (default) and type = “MIRT”, one plot per dimension will be generated.}

\item{same.plot}{If TRUE (default) and both \emph{H} and \emph{B} are supplied, generates both the Huber and bisquare plots in the same image frame. If FALSE, the Huber and bisquare plots are generated on separate images.}

\item{type}{Type of data: "Dichotomous" for dichotomous data (multidimensional or unidimensional) or "GRM" for Likert-type data}
}
\value{
‘Summary Statistics (Huber)’ If \emph{H} is supplied, a dataframe where each row provides a subject’s ID, MLE, the Huber-weighted robust estimate, the minimum distance between the point on the plot and the identity line, and their response vector. The subjects are organized by greatest to least distance.

‘Summary Statistics (Bisquare)’ If \emph{B} is supplied, a dataframe where each row provides a subject’s ID, MLE, the bisquare-weighted robust estimate, the minimum distance between the point on the plot and the identity line, and their response vector. The subjects are organized by greatest to least distance.

Plots If same.plot = TRUE and both \emph{H} and \emph{B} are supplied, each robust ability estimate is plotted against the MLE; graphs for each of the Huber- and bisquare-weighted estimates are generated separately but on the same image frame. The identity line \eqn{y=x} is plotted as a reference line.

`Huber Plot` If same.plot = FALSE or \emph{B} is not supplied, each Huber-weighted robust ability estimate is plotted against the MLE with the identity line \eqn{y=x} as reference.

`Bisquare Plot` If same.plot = FALSE or \emph{H} is not supplied, each bisquare-weighted robust ability estimate is plotted against the MLE with the identity line \eqn{y=x} as reference.
}
\description{
Generate a scatterplot of robust estimates versus the maximum likelihood estimate (MLE)
}
\details{
When the data is not disturbed, robust estimates should not differ greatly from the maximum likelihood estimate (MLE).
                                      By plotting the robust estimates against the MLE, the user can identify possible aberrant trends, if the robust estimates are far from the MLE, as indicated by the distance from the \eqn{y=x} identity line.
                                      Larger discrepancies between the point plotted for a subject and the identity line suggest there may be some disturbance in this subject’s data that the robust estimation may be correcting.
                                      At least one tuning parameter \emph{H} or \emph{B} must be supplied to the function; if both are supplied, the function will return separate plots for both weighting systems.
}
\examples{
## Test length
n <- 30

## Number of iterations of newton's method
iter <- 15

## Number of thresholds
nthresh <- 4

## Generate real thetas
thetas <- seq(-2, 2, by=.1)

## Generate item slope
a <- runif(n, .90, 2.15)

## Generate category threshold parameters
b <- matrix(runif(n*nthresh, -2.5,2.5), nrow = n, ncol =nthresh)
b <- t(apply(b, 1, sort))

## Calculate probabilities
probs <- item.prob(thetas, "GRM", cbind(a, b))

## Generate input data from probabilities
abdat <- data.gen(probs$P)

## Introduce aberrant responses: random guessing for latter 40\% of the exam
chng.pt <- .6
abdat[(chng.pt*n+1):n, ] <- sample(c(1:(nthresh+1)), length(thetas)*(n-chng.pt*n), replace = T)

## Plot the GRM
out<-theta_plots(abdat, a, b=b, iter=30, cutoff=0.01, H=.1, B=1, same.plot = F, type="GRM")
## Check bisquare plot
out$`Bisquare Plot`
## Check Huber summary
out$`Summary Statistics (Huber)`
}
