% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robust.R
\name{bisquare}
\alias{bisquare}
\title{Bisquare Weighting Function}
\usage{
bisquare(r, B)
}
\arguments{
\item{r}{Residual that measures the inconsistency of a response from the subject's assumed response model}

\item{B}{Bisquare tuning parameters}
}
\value{
Bisquare weighting function
}
\description{
This function return the value of bisquare weighting value given Residual and bisquare tuning paramter
}
