% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robust.R
\name{bisquare}
\alias{bisquare}
\title{Bisquare Weighting Function}
\usage{
bisquare(r, B)
}
\arguments{
\item{r}{A residual that measures the inconsistency of a response from the subject's assumed response model, on one item}

\item{B}{Bisquare tuning parameter}
}
\value{
Bisquare weight value
}
\description{
This function returns the value of the bisquare weight given a residual and bisquare tuning paramter
}
