% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robust.R
\name{choose.tuco}
\alias{choose.tuco}
\title{Plot histogram of residuals along plot of weight (dependent on TuCo) vs residuals}
\usage{
choose.tuco(r, H = NULL, B = NULL)
}
\arguments{
\item{r}{A vector of residuals}

\item{H}{Huber tuning parameter}

\item{B}{Bisquare tuning parameter}
}
\value{
Histogram plot of residuals beneath a graph of the weight functions vs. the residuals
}
\description{
This function plots a histogram of residuals along the graph of the weighting function (dependent on the tuning parameter) as a function of the residual
}
\details{
The goal of this plot is to visualize the proportion of residuals that are downweighted based on the tuning parameter and allow the researcher to choose a tuning parameter that suits their data well.
}
