% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robust.R
\name{choose.tuco}
\alias{choose.tuco}
\title{Plot histogram of residuals along plot of weight (dependent on TuCo) vs residuals}
\usage{
choose.tuco(r, H = NULL, B = NULL)
}
\arguments{
\item{r}{r is a vector of residuals}

\item{H}{Huber tuning parameter}

\item{B}{Bisquare tuning parameter}
}
\value{
Histogram plot
}
\description{
This function plots histogram of residuals along plot of weight (dependent on TuCo) vs residuals
}
\details{
<please add some details here>
}
