% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robust.R
\name{data.gen}
\alias{data.gen}
\title{Data generation function from probabilities of responding in each category}
\usage{
data.gen(P)
}
\arguments{
\item{H}{Huber tuning parameters}
}
\value{
The probabilities of responding in each category
}
\description{
This function generates data from probabilities of responding in each category
}
