% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robust.R
\name{huber}
\alias{huber}
\title{Huber Weighting Function}
\usage{
huber(r, H)
}
\arguments{
\item{r}{A residual that measures the inconsistency of a response from the subject's assumed response model, on one item}

\item{H}{Huber tuning parameter}
}
\value{
Huber weight value
}
\description{
This function returns the value of the Huber weight given a residual and Huber tuning paramter
}
