% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robust.R
\name{huber}
\alias{huber}
\title{Huber Weighting Function}
\usage{
huber(r, H)
}
\arguments{
\item{r}{Residual that measures the inconsistency of a response from the subject's assumed response model}

\item{H}{Huber tuning parameters}
}
\value{
Huber weighting function
}
\description{
This function return the value of Huber weighting value given Residual and Huber tuning paramter
}
