% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robust.R
\name{item.prob}
\alias{item.prob}
\title{Item Response Probability}
\usage{
item.prob(theta, model, ipars, D = 1.7)
}
\arguments{
\item{theta}{A numeric vector or matrix of latent trait values.}

\item{model}{A character string specifying which IRT model to use:
\itemize{
\item{\code{"Rasch}}{Allows item difficulty parameters to vary across items (Rasch, 1960).}
\item{\code{"1PL"}}{1‑parameter logistic model with a common discrimination parameter and item‑specific difficulties. The Rasch model is the special case where all discriminations equal 1.}
\item{\code{"2PL"}}{2‑parameter logistic model allowing both discrimination and difficulty to vary across items (Birnbaum, 1968).}
\item{\code{"MIRT"}}{Multidimensional extension of the 2PL model with item slope vectors across latent dimensions and item‑specific intercepts (McKinley & Reckase, 1983; Muraki & Engelhard, 1985).}
\item{\code{"GRM"}}{Graded response model for ordered polytomous items with item‑specific discrimination and ordered category thresholds (Samejima, 1969).}
\item{\code{"MGRM"}}{Multidimensional graded response model extending Samejima’s GRM to multiple latent dimensions, with slope vectors and ordered category thresholds (Muraki & Carlson, 1995).}
}}

\item{ipars}{A matrix of item parameters. See examples for how to structure the columns of the matrix based on the model utilized.}

\item{D}{A positive scaling constant used for scaling the normal ogive model. Defaults to 1.7; alternatively is often set to 1.0.}
}
\value{
For model accommodating dichotomous data ("1PL", "2PL", "MIRT"), returns an \eqn{N \times J} matrix of response probabilities \eqn{P(X = 1)}.

For models accommodating polytomous data ("GRM", "MGRM"), returns a list with: {pstar}: an array of cumulative probabilities \eqn{P^*(X \geq k)}.
        and {P}: an array of category probabilities \eqn{P(X = k)}.
}
\description{
Computes item response probabilities for select IRT models (1PL, Rasch, 2PL, MIRT, GRM, and MGRM), given ability and item parameters.
by constructing the appropriate linear predictors and applying the logistic function. Returns item response probabilities for dichotomous data or item category response probabilities for polytomous data.
}
\section{IRT Models}{


\strong{Rasch}

\deqn{
  P(X_{ij} = 1 \mid \theta_i) =
  \frac{1}{1 + e^{-(\theta_i - b_j)}}
}

\strong{1PL}

\deqn{
  P(X_{ij} = 1 \mid \theta_i) =
  \frac{1}{1 + e^{-Da(\theta_i - b_j)}}
}

\strong{2PL}

\deqn{
  P(X_{ij} = 1 \mid \theta_i) =
  \frac{1}{1 + e^{-Da_j(\theta_i - b_j)}}
}

\strong{MIRT}

\deqn{
  P(X_{ij} = 1 \mid \boldsymbol{\theta}_i) =
  \frac{1}{1 + e^{-D(\boldsymbol{a}_j\boldsymbol{\theta}_i + d_j)}}
}

\strong{GRM}

\deqn{
  P(X_{ij} = k \mid \theta_i) =
  \frac{1}{1 + e^{-Da_j(\theta_i - b_{jk})}}
  -
  \frac{1}{1 + e^{-Da_j(\theta_i - b_{j(k+1)})}}
}

\strong{MGRM}

\deqn{
  P(X_{ij} = k \mid \boldsymbol{\theta}_i) =
  \frac{1}{1 + e^{-D \sum_{l=1}^L a_{jl}(\boldsymbol{\theta}_{l} - d_{jk})}}
  -
  \frac{1}{1 + e^{-D \sum_{l=1}^L a_{jl}(\boldsymbol{\theta}_{l} - d_{j(k+1)})}}
}
}

\examples{
# Rasch case
N <- 3
J <- 5
theta <- rnorm(N) # generate ability values
ipars <- rnorm(J)  # generate item difficulties
item.prob(theta, "Rasch", ipars)
# 1PL case
N <- 4 # subjects
J <- 6 # items
theta <- rnorm(N) # generate ability values
ipars <- cbind(a = rep(1.2, J), # set item discrimination
               b = rnorm(J)) # generate item difficulties
item.prob(theta, "1PL", ipars)
# 2PL case
N <- 3 # subjects
J <- 5 # items
theta <- rnorm(N) # generate ability values
ipars <- cbind(a = runif(J, 0.5, 2), # generate item discrimination
               b = rnorm(J)) # generate item difficulty
item.prob(theta, "2PL", ipars) 
# MIRT case
N <- 2 # subjects
J <- 7 # items
L <- 3 # dimensions
theta <- matrix(rnorm(N * L), ncol = L) # N x L ability matrix
ipars <- cbind(matrix(runif(J * L, 0, 1), ncol = L), d = rnorm(J)) # generate slopes + intercept
item.prob(theta, "MIRT", ipars)
# GRM case
N <- 3 # subjects
J <- 5 # items
K <- 4 # categories
theta <- rnorm(N)
a <- runif(J, 0.5, 2) # J discriminations
b_raw <- matrix(rnorm(J * (K - 1)), nrow = J)  
b <- t(apply(b_raw, 1, sort)) # Sort thresholds within each item (row-wise)
ipars <- cbind(a, b) # Combine into a J x K matrix
item.prob(theta, "GRM", ipars)
# MGRM case
N <- 3 # subjects
J <- 5 # items
L <- 2 # dimensions
K <- 4 # categories
theta <- matrix(rnorm(N * L), ncol = L) 
a <- matrix(runif(J * L, 0, 1), ncol = L) # slopes 
b_raw <- matrix(rnorm(J * (K - 1)), nrow = J) # thresholds
b <- t(apply(b_raw, 1, sort))
ipars <- cbind(a, b)
item.prob(theta, "MGRM", ipars)
}
\references{
Birnbaum, A. (1968). Some latent trait models and their use in inferring an examinee’s ability. In F. M. Lord & M. R. Novick (Eds.), Statistical Theories of Mental Test Scores (pp. 397–479). \emph{Addison‑Wesley}.

Lord, F. M., & Novick, M. R. (1968). Statistical theories of mental test scores. \emph{Addison-Wesley}.

Lord, F. M. (1980). Applications of item response theory to practical testing problems. \emph{Erlbaum}. https://doi.org/10.4324/9780203056615.

McKinley, R. L., & Reckase, M. D. (1983). An extension of the two‑parameter logistic model to the multidimensional latent space. \emph{Psychometrika}, 48(3), 369–382.

Mosteller, F., & Tukey, J. W. (1977). Data Analysis and Regression: A Second Course in Statistics. \emph{Addison‑Wesley}.

Muraki, E., & Engelhard, G. (1985). Full-information item factor analysis: Applications of EAP scores. \emph{Applied Psychological Measurement}, 9(4), 417–430

Rasch, G. (1960). Probabilistic models for some intelligence and attainment tests. \emph{Danish Institute for Educational Research}, 184.

Samejima, F. (1969). Estimation of latent ability using a response pattern of graded scores. \emph{Psychometrika Monograph Supplement, 34} (4, Pt. 2), 100–100.
}
