% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robust.R
\name{probs.gen}
\alias{probs.gen}
\title{Response Probability Calculation (2PL)}
\usage{
probs.gen(thetas, a, d)
}
\arguments{
\item{thetas}{A vector of length p containing students’ latent traits, where p is the number of test dimensions}

\item{a}{A p x n matrix containing fixed item slope parameters}

\item{d}{A vector of length n containing fixed intercept parameters, for n test items}
}
\value{
A vector of probabilities of endorsing each of n test items
}
\description{
This function calculates the probabilities of getting items correct given person's abilities and item parameters
}
