% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robust.R
\name{probs.gen}
\alias{probs.gen}
\title{Response Probability Calculation (2PL)}
\usage{
probs.gen(thetas, a, d)
}
\arguments{
\item{thetas}{An array of person abilities}

\item{a}{A matrix of slope parameters}

\item{d}{Array of intercept parameters}
}
\value{
Array of probabilities of having correct response
}
\description{
This function calculates the probabilities of get items correct given person's abilities and item parameters
}
