% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robust.R
\name{pstar_to_p}
\alias{pstar_to_p}
\title{Covert p-star threshold values to the probabilities of responding in each category}
\usage{
pstar_to_p(pstar)
}
\arguments{
\item{pstar}{A matrix of p-star threshold values}
}
\value{
The probabilities of responding in each category
}
\description{
This function turns matrix of p-star threshold values into probabilities of responding in each category
}
