% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robust.R
\name{robust.item}
\alias{robust.item}
\title{Robust Estimation of Item Parameters}
\usage{
robust.item(dat)
}
\arguments{
\item{dat}{A \eqn{N\times J} matrix of dichotomous response data}
}
\description{
Robust Estimation of Item Parameters
}
\examples{
# Load package and example data set 
library(mirt) 
data(Science) 
# Robust estimation of item parameters
robust.item(Science)
}
\references{
Hong, M., & Cheng, Y. (2019). Robust maximum marginal likelihood (RMML) estimation for item response theory models. Behavior Research Methods, 51(2), 573–588. https://doi.org/10.3758/s13428-018-1150-4
}
