% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robust.R
\name{standard.errors}
\alias{standard.errors}
\title{Standard error function for Rasch, 1PL, 2PL, MIRT, GRM, MGRM}
\usage{
standard.errors(
  theta,
  ipars,
  dat,
  model,
  D = 1.7,
  weight.type = "equal",
  tuning.par = NULL,
  est.type = "MLE",
  prior = c(0, 1),
  eap.quad.pts = seq(-4, 4, length.out = 41)
)
}
\description{
Computes per person: information SE (expected Fisher), asymptotic SE (robust expected, incorporating weights), sandwich SE, Bayesian SE & bayesian sandwich SE (2PL & GRM only for Bayesian)
}
