% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robust.R
\name{theta.est.grm}
\alias{theta.est.grm}
\title{Ability Estimation Function Using Robust Estimation (GRM)}
\usage{
theta.est.grm(
  dat,
  a,
  b,
  iter = 30,
  cutoff = 0.01,
  init.val = 0,
  weight.type = "equal",
  tuning.par = NULL
)
}
\arguments{
\item{dat}{A \eqn{n \times m} matrix of polytomously-scored data (e.g., Likert-type) for \emph{n} items and \emph{m} subjects.}

\item{a}{Vector of slope parameters for \emph{n} items}

\item{b}{A \eqn{n \times (K-1)} matrix of category threshold parameters for \emph{K} categories}

\item{iter}{Max number of iterations. Default is 100}

\item{cutoff}{Threshold value to terminate the iteration when the likelihood changes below this value, which means that the estimation is converged.}

\item{init.val}{Vector of initial latent trait for the maximum likelihood estimation for \emph{m} subjects. If a single value is provided, that initial value will be used for all subjects. Default is 0.}

\item{tuning.par}{The tuning parameter for "bisquare" or "Huber" weighting functions. Greater tuning parameters result in less downweighting in robust estimation.}

\item{weight.category}{The weighting strategy to use, "equal", "bisquare" and "Huber". Default is "equal", which is equally weighted as in standard maximum likelihood estimation.}
}
\value{
theta Ability estimates for \emph{m} subjects. NAs replace values that did not converge to any value. Estimates that converged to values less than -3.0 were replaced with -3.0, while estimates that converged to values greater than 3.0 were replaced with 3.0.

convergence Indicators of convergence for \emph{m} subjects: a “0” indicates the value converged, while a “1” indicates the maximum likelihood estimation did not converge to any value.

theta.progression A matrix with rows corresponding to each subject and columns corresponding to the number of iterations supplied to the input. Each column provides the updated theta estimate at each iteration of the Newton-Raphson algorithm until the change in log-likelihood for that subject reaches the cutoff value or the value is nonconverged (reaches infinite values).

residual A \eqn{n \times m \times iter} array containing residuals corresponding to the ability estimate for \emph{m} subjects respective to the \emph{n} test items at each iteration until convergence, nonconvergence, or singular matrix is reached.
}
\description{
This function returns ability estimations based response data, item parameters for the graded respose model, and the given weighting function
}
\details{
The goal of robust estimation is to downweigh potentially aberrant responses to lessen their impact on the estimation of \eqn{\theta}. Robust estimates resist the harmful effects of response disturbances and tend to be less biased estimates of true ability than maximum likelihood estimates.
Under the graded response model (GRM), the contribution of item \emph{i} to the overall log-likelihood for one subject is weighted with a weight \eqn{\omega(r_i)} as a function of a residual \eqn{r_i} for the item:
\deqn{\sum^n_{i=1} \omega(r_i) \sum^K_{k=1} u_{ik}\text{log}P_{ik} = 0 }
The residual, which measures the inconsistency of a response from the subject's assumed response model, is \deqn{r_i = \frac{1}{K-1}\left[X_i - E(X_i|\hat{\theta})\right]} for the GRM. 
Two types of weight functions are used: Tukey's bisquare weighting function (Mosteller & Tukey, 1977)
\deqn{\omega(r_i)=\begin{cases}[1-(r_i/B)^2]^2, & \text{if} |r_i|\leq B.\\0, & \text{if} |r_i|>B.\end{cases}}
and the Huber weighting function (Huber, 1981)
\deqn{\omega(r_i)=\begin{cases}1, & \text{if} |r_i|\leq H.\\H/|r_i|, & \text{if} |r_i|>H.\end{cases}}
Both functions are effective in estimating more accurate scores with aberrant data, although the bisquare weight function may lead to nonconvergence when using data containing a high proportion of incorrect responses (Schuster & Yuan, 2011).
}
\references{
Huber, P. (1981) \emph{Robust Statistics}. Wiley, New York. https://doi.org/10.1002/0471725250

Mosteller, F., & Tukey, J. W. (1977). \emph{Data Analysis and Regression: A Second Course in Statistics}. Reading, MA: Addison-Wesley Pub Co.

Schuster, C., & Yuan, K.-H. (2011). Robust Estimation of Latent Ability in Item Response Models. \emph{Journal of Educational and Behavioral Statistics}, 36(6), 720–735. https://doi.org/10.3102/1076998610396890
}
