% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robust.R
\name{theta.est.grm}
\alias{theta.est.grm}
\title{Ability Estimation Function Using Robust Estimation (GRM)}
\usage{
theta.est.grm(
  dat,
  a,
  b,
  iter = 30,
  cutoff = 0.01,
  init.val = 0,
  weight.type = "equal",
  tuning.par = NULL,
  D = 1.7
)
}
\arguments{
\item{dat}{A \eqn{J \times N} matrix of polytomously-scored data (e.g., Likert-type) for \emph{J} items and \emph{N} subjects.}

\item{a}{Vector of slope parameters for \emph{J} items}

\item{b}{A \eqn{J \times (K-1)} matrix of category threshold parameters for \emph{K} categories}

\item{iter}{Max number of iterations. Default is 100}

\item{cutoff}{Threshold value to terminate the iteration when the likelihood changes below this value, which means that the estimation is converged.}

\item{init.val}{Vector of initial latent trait for the maximum likelihood estimation for \emph{N} subjects. If a single value is provided, that initial value will be used for all subjects. Default is 0.}

\item{tuning.par}{The tuning parameter for "bisquare" or "Huber" weighting functions. Greater tuning parameters result in less downweighting in robust estimation.}

\item{weight.category}{The weighting strategy to use: "equal", "bisquare" and "Huber". Default is "equal", which is equally weighted as in standard maximum likelihood estimation.}
}
\value{
theta Ability estimates for \emph{N} subjects. NAs replace values that did not converge to any value. Estimates that converged to values less than -3.0 were replaced with -3.0, while estimates that converged to values greater than 3.0 were replaced with 3.0.

convergence Indicators of convergence for \emph{N} subjects: a “0” indicates the value converged, while a “1” indicates the maximum likelihood estimation did not converge to any value.

standard.error Standard errors of the theta estimates for \emph{N} subjects, given by the square root of the reciprocal of the Fisher information. NAs replace nonconverging values.

theta.progression A matrix with rows corresponding to each subject and columns corresponding to the number of iterations supplied to the input. Each column provides the updated theta estimate at each iteration of the Newton-Raphson algorithm until the change in log-likelihood for that subject reaches the cutoff value or the value is nonconverged (reaches infinite values).

residual A \eqn{J \times N \times p} array containing residuals corresponding to the ability estimate for \emph{N} subjects respective to the \emph{J} test items at each iteration until convergence within maximum \emph{p} iterations, nonconvergence, or singular matrix is reached.
}
\description{
Calculate robust ability estimates using the GRM item response function with the given weight function, fixed item parameters, and item responses
}
\details{
The goal of robust estimation is to downweigh potentially aberrant responses to lessen their impact on the estimation of \eqn{\theta_i}. Robust estimates resist the harmful effects of response disturbances and tend to be less biased estimates of true ability than maximum likelihood estimates.
              Under the graded response model (GRM; Samejima, 1969), the probability that a subject responds in or above a category \emph{k} for item \emph{j} is \eqn{P^*_{jk}(\theta_i) = \frac{1}{1+ e^{-a_j (\theta_i-b_{jk})}}}  (Embretson & Reise, 2000). \eqn{a_j} is the item discrimination parameter. There are \emph{K} categories and \eqn{K-1} threshold parameters (\eqn{b_{j,1}, ..., b_{j,K-1}}), where \eqn{b_{j,k}} separates response category \eqn{k} and \eqn{k+1} (\eqn{k=1,..K-1}).
              The probability of endorsing exactly category \eqn{k} is therefore: \eqn{P_{jk}(\theta_i) = P^*_{j,k}(\theta_i) - P^*_{j,k+1}(\theta_i),} where \eqn{P^*_{j1}(\theta_i) \equiv 1.0} and \eqn{P^*_{jK}(\theta_i) \equiv 0.0.}
              The contribution of item \emph{j} to the overall log-likelihood for one subject is weighted with a weight \eqn{\omega(r_{ij})} as a function of a residual \eqn{r_{ij}} for the item:
              \deqn{\sum^J_{j=1} \omega(r_{ij}) \sum^K_{k=1} u_{jk}\text{log}P_{jk} = 0 }
              \eqn{u_{jk}} is an indicator function: \deqn{u_{jk} = \begin{cases}
                                                           1 & \text{if } X_{ij} = k; \\
                                                           0 & \text{otherwise}.
                                                           \end{cases} }
              The residual, which measures the inconsistency of a response from the subject's assumed response model, is \deqn{r_{ij} = \frac{1}{\sigma_{X_{ij}}}\left[X_{ij} - E(X_{ij}|\hat{\theta}_i)\right]} for the GRM.
              The difference in fit is determined between the observed response \eqn{X_{ij}} and expected score \eqn{E(X_{ij}|\hat{\theta}_i) = \sum_{k=1}^KkP_{jk}(\hat{\theta}_i)}, and scaled by the variance \eqn{\sigma_{X_{ij}}^2 = \sum_{k=1}^K (X_{ijk}-E[X_{ij}|\hat{\theta}_i])^2P_{jk}(\hat{\theta}_i).}
              Two types of weight functions are used: Tukey's bisquare weighting function (Mosteller & Tukey, 1977)
                \deqn{\omega(r_{ij})=\begin{cases}[1-(r_{ij}/B)^2]^2, & \text{if} |r_{ij}|\leq B.\\0, & \text{if} |r_{ij}|>B.\end{cases}}
              and the Huber weighting function (Huber, 1981)
                \deqn{\omega(r_{ij})=\begin{cases}1, & \text{if} |r_{ij}|\leq H.\\H/|r_{ij}|, & \text{if} |r_{ij}|>H.\end{cases}}
              Both functions are effective in estimating more accurate scores with aberrant data, although the bisquare weight function may lead to nonconvergence when using data containing a high proportion of incorrect responses (Schuster & Yuan, 2011).
}
\examples{
# Test Length
n<-30

# Number of thresholds (5-point Likert scale)
nthresh<-4

# Number of iterations of Newton's method
iter <- 15

# Set critical value for convergence criteria
crit.val<-0.01

# Set real thetas - 5 subjects
thetas<-c(-2,-1,0,1,2)

# Set item slope
a<-runif(n, .90, 2.15)

# Set threshold parameters
b<-t(apply(matrix(runif(n*4, -2.5,2.5), nrow = n, ncol =4), 1, sort))

# Calculate Probabilities
probs<-item.prob(thetas, "GRM", cbind(a, b))

# Generate Likert data
dat<-data.gen(probs$P)

# Make the data aberrant by reverse coding 20\% items
ab.prop<-0.2
index<-sample(c(1:n), ab.prop*n)
ab.dat<-dat
ab.dat[index, ]<-apply(matrix(dat[index,]), c(1,2), function(x) return(nthresh+2-x))


# Calculate MLE (Non-robust)
mle<-theta.est.grm(ab.dat, a, b, iter, crit.val, init.val=0, weight.type="equal")

# Use MLE as starting value, or 0 if NA
start.val<-apply(mle$theta, c(1,2), function(x) ifelse(is.na(x), 0, x))

# Calculate bisquare- and Huber-weighted robust estimates
b.est<- theta.est.grm(ab.dat, a, b, iter, crit.val, init.val=start.val, weight.type="bisquare", tuning.par=4)
h.est<-theta.est.grm(ab.dat, a, b, iter, crit.val, init.val=start.val, weight.type="Huber", tuning.par=1)

# Compare robust ability estimates with MLE
b.est$theta
h.est$theta
mle$theta

}
\references{
Embretson, S. E., & Reise, S. P. (2000). \emph{Item response theory for psychologists.} Mahwah, N.J: L. Erlbaum Associates.

Huber, P. (1981) \emph{Robust Statistics}. Wiley, New York. https://doi.org/10.1002/0471725250

Mosteller, F., & Tukey, J. W. (1977). \emph{Data Analysis and Regression: A Second Course in Statistics}. Reading, MA: Addison-Wesley Pub Co.

Samejima, F. (1969). Estimation of latent ability using a response pattern of graded scores. \emph{Psychometrika Monograph Supplement, 34} (4, Pt. 2), 100–100.

Schuster, C., & Yuan, K.-H. (2011). Robust Estimation of Latent Ability in Item Response Models. \emph{Journal of Educational and Behavioral Statistics}, 36(6), 720–735. https://doi.org/10.3102/1076998610396890
}
