% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robust.R
\name{theta.est}
\alias{theta.est}
\title{Ability Estimation Function Using Robust Estimation}
\usage{
theta.est(
  dat,
  a,
  d,
  iter = 100,
  cutoff = 0.01,
  theta.initial = rep(0, ncol(a)),
  weight.type = "equal",
  tuning.par = NULL
)
}
\arguments{
\item{dat}{Response data to load}

\item{a}{Matrix of slope parameters}

\item{d}{Array of intercept parameters}

\item{iter}{Max number of iterations. Default is 100}

\item{cutoff}{Threshold value to terminate the iteration when the likelihood changes below this value, which means that the estimation is converged.}

\item{theta.initial}{Array of initial thetas. Default is 0s}

\item{tuning.par}{The tuning parameter for "bisquare" or "Huber" weighting functions}

\item{weight.category}{The weighting strategy to use, "equal", "bisquare" and "Huber". Default is "equal", which is equally weighted.}
}
\value{
Array of estimated person abilities
}
\description{
This function return the list of ability estimations based on the given weighting function
}
\details{
Bisquqre weighting function
\deqn{\omega(r_i)=\begin{cases}[1-(r_i/B)^2]^2, & \text{if} |r_i|\leq B.\\0, & \text{if} |r_i|>B.\end{cases}}
Huber weighting function
\deqn{\omega(r_i)=\begin{cases}1, & \text{if} |r_i|\leq H.\\H/|r_i|, & \text{if} |r_i|>H.\end{cases}}
where \eqn{r_i} measures the inconsistency of a response of item i from the subject's assumed response model. Mislevy and Bock proposed a residual for the unidimentional case, \deqn{r_i=a_i(\theta-b_i)}
}
