% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robust.R
\name{theta.est}
\alias{theta.est}
\title{Ability Estimation Function Using Robust Estimation - Dichotomous Data}
\usage{
theta.est(
  dat,
  a,
  d,
  D = 1.7,
  iter = 30,
  cutoff = 0.01,
  init.val = rep(0, ncol(a)),
  weight.type = "equal",
  tuning.par = NULL,
  residual = "information"
)
}
\arguments{
\item{dat}{A \eqn{J \times N} matrix of dichotomously-coded data where 1 represents an endorsed response and 0 represents a non-endorsed response. \emph{J} is the test length and \emph{N} is the number of subjects.}

\item{a}{A \eqn{L \times J} matrix containing fixed item slope parameters for \emph{J} items and \emph{L} dimensions}

\item{d}{A vector of length \emph{J} containing fixed intercept parameters for J items. If the model is unidimensional, this vector should contain difficulty parameters corresponding to \emph{b} in the unidimensional 2PL model.}

\item{iter}{Maximum number of iterations for the Newton-Raphson method. Default is 30.}

\item{cutoff}{Threshold value to terminate the iteration when the likelihood changes below this value, which means that the estimation is converged. Default is 0.01.}

\item{init.val}{A vector of length \emph{L} containing initial latent trait values for the maximum likelihood estimation. The default is 0 for each of the \emph{L} dimensions.}

\item{tuning.par}{The tuning parameter for "bisquare" or "Huber" weighting functions. Greater tuning parameters result in less downweighting.}

\item{residual}{Type of residual if using bisquare or Huber weight functions. Default is "information" while "standardized" can alternatively be used.}

\item{weight.category}{The weighting strategy to use: "equal", "bisquare", or "Huber". Default is "equal", which is equally weighted as in standard maximum likelihood estimation.}
}
\value{
theta A \eqn{N \times L} matrix of ability estimates for \emph{N} subjects and \emph{L} dimensions.

standard.errors A \eqn{N \times L} matrix of the standard errors of ability estimates for \emph{N} subjects and \emph{L} dimensions, calculated by the square root of the reciprocal of the Fisher information. NAs replace nonconverging values.

convergence \eqn{N \times L} matrix containing indicators of convergence for \emph{N} subjects: a “0” indicates the value converged; a “1” indicates the maximum likelihood estimation did not converge to any value; and “Singular” indicates the Hessian matrix was singular and could not be used to continue the maximum likelihood estimation.

theta.progression A \eqn{L \times p \times N} array for \emph{L} dimensions, \emph{p} number of iterations supplied to the input, and \emph{N} subjects. Each column provides the updated theta estimate at each iteration of the Newton-Raphson algorithm until the change in log-likelihood for that subject reaches the cutoff value, infinite values (nonconverged), or encounters a singular matrix error.

residual A \eqn{J \times N} matrix with residuals corresponding to the ability estimate for \emph{N} subjects respective to the \emph{J} test items.
}
\description{
Calculate robust ability estimates using the MIRT item response function with the given weight function, fixed item parameters, and item responses
}
\details{
The goal of robust estimation is to downweigh potentially aberrant responses to lessen their impact on the estimation of \eqn{\boldsymbol{\theta}}. Robust estimates resist the harmful effects of response disturbances and tend to be less biased estimates of true ability than maximum likelihood estimates.
              Using the multidimensional IRT model for dichotomous data (McKinley & Reckase, 1983), the probability of a correct response (e.g., "1") on item \emph{j} is given by the formula \eqn{P(X_{ij}=1 | \boldsymbol{\theta}_i) =  \frac{1}{1+ e^{-1.7(\boldsymbol{a}_j\boldsymbol\theta_i+d_j)}},} where \eqn{\boldsymbol{a}_j} is a \eqn{L \times J} matrix of item slope parameters for \emph{J} items and \emph{L} dimensions. The intercept parameter is \eqn{d_j}.
}
\examples{
# Test length
n <- 30

## Number of iterations of Newton's method
iters<-50

## Number of dimensions
dim <- 3

## Correlation between one person's thetas
cor <- .6

## Covariance matrix for generating thetas
sigma <- matrix(cor, ncol = dim, nrow = dim)
diag(sigma) <- 1
s <- rep(1, dim)
sigma <- s*sigma*s

## Generate real thetas
thetas <- matrix(mvrnorm(n = 10, rep(0,dim), Sigma = sigma), ncol = dim)

## Generate slope parameters
a <- matrix(runif(n*dim, 1.5, 3), nrow = n, ncol = dim)
## Generate intercept parameters
d <- matrix(rnorm(n), ncol = 1)
## Calculate probabilities
probs <- item.prob(thetas, "MIRT", cbind(a,d))
## Generate data from probabilities
dat <- apply(probs, c(1,2), function(x) rbinom(1,1,x))

## Estimate thetas
theta.est(dat, a, d, iters, init.val=matrix(rep(0,dim)), weight.type="Huber", tuning.par=1)
}
\references{
Huber, P. (1981) \emph{Robust Statistics}. Wiley, New York. https://doi.org/10.1002/0471725250.

McKinley, R. L., & Reckase, M. D. (1983, August). \emph{An Extension of the Two-Parameter Logistic Model to the Multidimensional Latent Space} (Research No. ONR83-2). Iowa City, IA: American College Testing Program.

Mosteller, F., & Tukey, J. W. (1977). \emph{Data Analysis and Regression: A Second Course in Statistics}. Reading, MA: Addison-Wesley Pub Co.

Schuster, C., & Yuan, K.-H. (2011). Robust Estimation of Latent Ability in Item Response Models. \emph{Journal of Educational and Behavioral Statistics}, 36(6), 720–735. https://doi.org/10.3102/1076998610396890.
}
