% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robust.R
\name{theta_plots}
\alias{theta_plots}
\title{Plot to compare robust estimates with MLE}
\usage{
theta_plots(
  dat,
  a,
  d,
  iter = 30,
  cutoff = 0.01,
  H = NULL,
  B = NULL,
  same.plot.dim = F,
  same.plot = T,
  type
)
}
\arguments{
\item{dat}{Response data to load}

\item{a}{Matrix of slope parameters}

\item{iter}{Max number of iterations. Default is 100}

\item{cutoff}{Threshold value to terminate the iteration when the likelihood changes below this value, which means that the estimation is converged.}

\item{H}{Huber tuning parameter}

\item{B}{Bisquare tuning parameter}

\item{same.plot.dim}{If in same plot}

\item{same.plot}{If allow user to have sperate plots or see both Huber and Bisquare at same time}

\item{type}{MIRT or GRM}

\item{b}{Array of intercept parameters}
}
\value{
Histogram plot
}
\description{
This function plots Robust Estimates vs the MLE
}
\details{
<please add some details here>
}
